Move AssertionExtensions to shared DBTests project; relocate OdsApiValidatorTests to Common.UnitTests
