+++ Application/EdFi.Ods.AdminApi.DBTests.Common/EdFi.Ods.AdminApi.DBTests.Common.csproj
@@ -10,5 +10,7 @@
     <PackageReference Include="Respawn" />
     <PackageReference Include="EdFi.Suite3.Admin.DataAccess" />
     <PackageReference Include="EdFi.Suite3.Security.DataAccess" />
+    <PackageReference Include="FluentValidation.AspNetCore" />
+    <PackageReference Include="Shouldly" />
   </ItemGroup>
 </Project>
